+++ refact-agent/gui/src/services/graphql/index.ts
@@ -1,0 +1,3 @@
+export * from "./queriesAndMutationsApi";
+export * from "./subscriptions";
+export * from "./actions";
